Refactor: Factor out database creation/deletion classes into tests.
